Round-Shop cycle

+++ Assets/Scripts/RunProgress/Round.cs
@@ -9,7 +9,7 @@ namespace RunProgress
     {
         [field: SerializeField] public EnemySpawnCycle SpawnCycle { get; private set; }
         [field: SerializeField] public RoundWeaponAction WeaponAction { get; private set; }
-        [field: SerializeField] [Tooltip("In seconds")] public int Length { get; private set; }
+        [field: SerializeField] [Tooltip("In seconds")] public float Length { get; private set; }
     }
 
     public enum RoundWeaponAction
